refactor(tab-manager): consolidate schema into workspace, deduplicate definition
Flatten schema/ folder into workspace.ts as the single source of truth
for table definitions, workspace definition, and types. Extract popup
client to workspace-popup.ts. Move row-converters to sync/ folder since
they're sync logic, not schema. Remove dead BrowserDb type alias.
